feat: allow default arguments for options in the command line
This allows you to specify default values for options that are present
in the command line command, but have no parameters specified. This
allows us to distinguish between having an option specified, but with
no parameters, so it uses the default value, and having that option not
being specified at all, using either the default value, or not using
anything at all.

Signed-off-by: flow <flowlnlnln@gmail.com>