fix recursive crash on win64
one of the contributors to PolyMC that moved over to Prism (TheLastRar) ended up finding this.
I didn't understand what it was for until I fixed & ran a 64-bit Windows build myself, and got a 0xC00000FD (Stack overflow).
This is caused by the massive recursive load of ATLauncher packs.
This is a pretty jank fix as it is, and later should be re-engineered to not be recursive.

Signed-off-by: jdp_ <42700985+jdpatdiscord@users.noreply.github.com>