Add "Create desktop shortcut" feature.
This patch adds a way of adding a shortcut to directly start an
instance. How this works is it uses the already existing "-l" command
line option. This, however, has not been implemented for Mac, which is
something that can definetly be improved in the future if anyone wants
to work on it :^)

Signed-off-by: xSlendiX <slendi@socopon.com>