Create offline mode accounts with consistent Uuids
Uses basically what already exists in Qt, just modified so I can specify
an arbitrary length byte array for the namespace. Not really sure if it
needs to be the same as the way Minecraft generates it or if I could've
just used Uuid::createUuidV3. It would've still been consistent either
way, but whatever, this works.

Signed-off-by: Lenny McLennington <lenny@sneed.church>