Skip non-executable Java paths when probing
Instead of blindly spawning a potentially excessive number of processes,
nonexistent or non-executable paths could be skipped by checking the
path returned by FS::ResolveExecutable. If it returns an empty string,
the path does not resolve to an executable file and can be skipped.

On Linux, this change dramatically speeds up the Java auto-detection.

Signed-off-by: byte-chan™ <baumi@bytesoft.dev>