Disable __GL_THREADED_OPTIMIZATIONS on discrete graphics
Fixes minecraft freezes/crashes when full-screening/exiting the game. The problem exists for like a year on all my devices and it also was reported on nvidia/reddit forums by users. On old versions, like 1.8.9 this happens every minecraft launch. On new versions I don't see the problem, but here might be some related artefacts (black line in bottom on random minecraft launch). The solution was found by a user on nvidia forum, but it seems, like no fix was applied: https://forums.developer.nvidia.com/t/opengl-crash-when-entering-fullscreen-on-rtx-3060-mobile-with-nvidia-dkms-555-58-02-minecraft/299192/26
This commit implements a workaround in PolyMC useDiscreteGpu option.

Signed-off-by: Vadim Vetrov <vetrovvd@gmail.com>